Fixed min() and max() handling arrays of integers and numbers.
