+++ src/jmespath.net/Functions/MathArrayFunction.cs
@@ -1,3 +1,8 @@
+using DevLab.JmesPath.Utils;
+using Newtonsoft.Json.Linq;
+using System.Collections.Generic;
+using System.Linq;
+
 namespace DevLab.JmesPath.Functions
 {
     public abstract class MathArrayFunction : JmesPathFunction
@@ -17,5 +22,20 @@ namespace DevLab.JmesPath.Functions
             base.Validate();
             EnsureArrayOfSame(args[0], "number", "string");
         }
+
+        public override JToken Execute(params JmesPathFunctionArgument[] args)
+        {
+            var array = (JArray) args[0].Token;
+
+            if (array.Count == 0)
+                return JTokens.Null;
+
+            var item = array[0];
+            var dataTypes = array.Select(t => t.GetTokenType()).Distinct();
+
+            return Execute(array, dataTypes);
+        }
+
+        protected abstract JToken Execute(JArray array, IEnumerable<string> dataTypes);
     }
 }
