+++ src/jmespath.net/Functions/MaxFunction.cs
@@ -1,48 +1,16 @@
-using System;
 using System.Linq;
-using DevLab.JmesPath.Utils;
 using Newtonsoft.Json.Linq;
 
 namespace DevLab.JmesPath.Functions
 {
-    public class MaxFunction : MathArrayFunction
+    public class MaxFunction : MinOrMaxFunction
     {
         public MaxFunction()
-            : base("max", 1)
+            : base("max")
         {
         }
 
-        public override JToken Execute(params JmesPathFunctionArgument[] args)
-        {
-            var array = (JArray)args[0].Token;
-
-            if (array.Count == 0)
-                return JTokens.Null;
-
-            var item = array[0];
-            var type = item.GetTokenType();
-
-            switch (type)
-            {
-                case "number":
-                    {
-                        if (item.Type == JTokenType.Float)
-                            return GetMax<double>(array);
-
-                        else /* if (token.Type == JTokenType.Integer) */
-                            return GetMax<long>(array);
-                    }
-
-                case "string":
-                    return GetMax<string>(array);
-
-                default:
-                    System.Diagnostics.Debug.Assert(false);
-                    throw new NotSupportedException("Error: invalid-type");
-            }
-        }
-
-        private static JToken GetMax<T>(JArray array)
+        protected override JToken GetMinOrMax<T>(JArray array)
         {
             var sequence = array
                 .Select(u => u.Value<T>())
